+++ R/deployApp.R
@@ -4,6 +4,20 @@
 #' [RMarkdown][rmarkdown::rmarkdown-package] document, a plumber API, or HTML
 #' content to a server.
 #'
+#' ## Updating existing apps
+#'
+#' If you have previously deployed an app, `deployApp()` will do its best to
+#' update the existing deployment. In the simple case where you have only one
+#' deployment in `appDir`, this should just work with the default arguments.
+#' If you want multiple deployments to the same server, supply `appName`.
+#' If you want multiple deployments to different servers, use `account` and/or
+#' `server`.
+#'
+#' The metadata needs to make this work is stored in the `rsconnect/` directory
+#' beneath `appDir`. You should generally check these files into version
+#' control to ensure that future you and other collaborators will publish
+#' to the same location.
+#'
 #' @param appDir Directory containing application. Defaults to current working
 #'   directory.
 #' @param appFiles The files and directories to bundle and deploy (only if
@@ -23,20 +37,22 @@
 #'   qualified path. Deployment information returned by [deployments()] is
 #'   associated with the source document.
 #' @param appName Name of application (names must be unique within an account).
-#'   Defaults to the base name of the specified `appDir`.
+#'   If not supplied on the first deploy, it will be generated from the base
+#'   name of `appDir` and `appTitle`, if supplied. On subsequent deploys,
+#'   it will use the previously stored value.
 #' @param appTitle Free-form descriptive title of application. Optional; if
-#'   supplied, will often be displayed in favor of the name. When deploying a
-#'   new application, you may supply only the `appTitle` to receive an
-#'   auto-generated `appName`.
+#'   supplied, will often be displayed in favor of the name. If ommitted,
+#'   on second and subsequent deploys, the title will be unchanged.
 #' @param appId If updating an application, the ID of the application being
-#'   updated. Optional unless updating an app owned by another user.
+#'   updated. For shinyapps.io, this the `id` listed on your applications
+#'   page. For Posit Connect, this is the `guid` that you can find on the
+#'   info tab on the content page. Generally, you should not need to supply
+#'   this as it will be automatically taken from the deployment record on disk.
 #' @param contentCategory Optional; the kind of content being deployed (e.g.
 #'   `"plot"` or `"site"`).
-#' @param account Account to deploy application to. This parameter is only
-#'   required for the initial deployment of an application when there are
-#'   multiple accounts configured on the system (see [accounts]).
-#' @param server Server name. Required only if you use the same account name on
-#'   multiple servers.
+#' @param account,server Uniquely identify a remote server with either your
+#'   user `account`, the `server` name, or both. Use [accounts()] to see the
+#'   full list of available options.
 #' @param upload If `TRUE` (the default) then the application is uploaded from
 #'   the local system prior to deployment. If `FALSE` then it is re-deployed
 #'   using the last version that was uploaded. `FALSE` is only supported on
@@ -608,7 +624,7 @@ getAppById <- function(id, account = NULL, server = NULL, hostUrl = NULL) {
   accountDetails <- NULL
   tryCatch({
     # attempt to look up the account locally
-    accountDetails <- accountInfo(resolveAccount(account, server), server)
+    accountDetails <- accountInfo(account, server)
   }, error = function(e) {
     # we'll retry below
   })