+++ tests/testthat/test-deploymentTarget.R
@@ -4,27 +4,17 @@ test_that("errors if no accounts", {
   expect_snapshot(deploymentTarget(), error = TRUE)
 })
 
-test_that("errors if unknown server", {
-  # TODO(HW): give better error for unknown server
-  expect_snapshot(deploymentTarget(server = "baz"), error = TRUE)
-})
-
-test_that("errors if bad account", {
-  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
+test_that("errors if unknown account or server", {
+  mockr::local_mock(accounts = fakeAccounts("foo", "bar"))
 
   expect_snapshot(error = TRUE, {
-    deploymentTarget(server = NULL, account = "john")
+    deploymentTarget(server = "unknown")
+    deploymentTarget(account = "john")
   })
 })
 
 test_that("succeeds if app is fully specified", {
-  mockr::local_mock(
-    accounts = fakeAccounts("ron", "bar"),
-    accountInfo = fakeAccountInfo(ron = list(
-      username = "foo",
-      server = "test"
-    ))
-  )
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
 
   app_dir <- withr::local_tempdir()
   file.create(file.path(app_dir, "app.R"))
@@ -37,7 +27,6 @@ test_that("succeeds if app is fully specified", {
     account = "ron"
   )
   expect_equal(target$appId, "123")
-  expect_equal(target$username, "foo")
 })
 
 test_that("errors if no previous deployments and multiple accounts", {
@@ -53,21 +42,15 @@ test_that("errors if no previous deployments and multiple accounts", {
 })
 
 test_that("handles accounts if only server specified", {
-  mockr::local_mock(
-    accounts = fakeAccounts(c("ron", "john"), "foo"),
-    accountInfo = fakeAccountInfo(ron = list(
-      username = "foo",
-      server = "test"
-    ))
-  )
+  mockr::local_mock(accounts = fakeAccounts(c("ron", "john"), "foo"))
 
   app_dir <- withr::local_tempdir()
   file.create(file.path(app_dir, "app.R"))
 
   expect_snapshot(deploymentTarget(app_dir, server = "foo"), error = TRUE)
 
-  target <- deploymentTarget(app_dir, server = "foo", account = "ron", appId = "123")
-  expect_equal(target$username, "foo")
+  target <- deploymentTarget(app_dir, server = "foo", account = "ron")
+  expect_equal(target$username, "ron")
 })
 
 test_that("errors if multiple deployments", {
@@ -77,24 +60,16 @@ test_that("errors if multiple deployments", {
   saveDeployment(
     app_dir,
     name = "test",
-    title = "",
-    username = "ron",
-    account = "ron",
     server = "foo1",
-    hostUrl = "",
-    appId = "123",
+    appId = "1",
     bundleId = "abc",
     url = "http://example.com"
   )
   saveDeployment(
     app_dir,
     name = "test",
-    title = "",
-    username = "ron",
-    account = "ron",
     server = "foo2",
-    hostUrl = "",
-    appId = "123",
+    appId = "2",
     bundleId = "abc",
     url = "http://example.com"
   )
@@ -106,54 +81,88 @@ test_that("errors if multiple deployments", {
 })
 
 test_that("succeeds if there's a single existing deployment", {
-  mockr::local_mock(
-    accounts = fakeAccounts("ron", "bar"),
-    accountInfo = fakeAccountInfo(ron = list(
-      username = "foo",
-      server = "test"
-    ))
-  )
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
 
   app_dir <- withr::local_tempdir()
   file.create(file.path(app_dir, "app.R"))
   saveDeployment(
     app_dir,
     name = "test",
-    title = "my title",
-    username = "ron",
-    account = "ron",
-    server = "bar",
-    hostUrl = "",
-    appId = "123",
+    appId = "1",
     bundleId = "abc",
     url = "http://example.com"
   )
 
-  target <- deploymentTarget(app_dir, appId = "123")
-  expect_equal(target$appId, "123")
-  expect_equal(target$appTitle, "my title")
+  target <- deploymentTarget(app_dir)
+  expect_equal(target$appId, "1")
+  expect_equal(target$username, "ron")
 
-  target <- deploymentTarget(app_dir, appName = "test", appId = "123")
-  expect_equal(target$appId, "123")
-  # TODO(HW): why doesn't this equal my title?
-  # expect_equal(target$appTitle, "my title")
+  target <- deploymentTarget(app_dir, appName = "test")
+  expect_equal(target$appId, "1")
+  expect_equal(target$username, "ron")
+})
+
+test_that("new title overrides existing title", {
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
+
+  app_dir <- withr::local_tempdir()
+  file.create(file.path(app_dir, "app.R"))
+  saveDeployment(
+    app_dir,
+    name = "test",
+    title = "old title",
+    appId = "1",
+    bundleId = "abc",
+    url = "http://example.com"
+  )
+
+  target <- deploymentTarget(app_dir)
+  expect_equal(target$appTitle, "old title")
+
+  target <- deploymentTarget(app_dir, appTitle = "new title")
+  expect_equal(target$appTitle, "new title")
 })
 
 test_that("succeeds if there are no deployments and a single account", {
-  mockr::local_mock(
-    accounts = fakeAccounts("ron", "bar"),
-    accountInfo = fakeAccountInfo(ron = list(
-      username = "ron_smith",
-      server = "test"
-    ))
-  )
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
 
   app_dir <- withr::local_tempdir()
   file.create(file.path(app_dir, "app.R"))
 
-  target <- deploymentTarget(app_dir, appId = "123")
-  expect_equal(target$username, "ron_smith")
+  target <- deploymentTarget(app_dir)
+  expect_equal(target$username, "ron")
 
-  target <- deploymentTarget(app_dir, appName = "foo", appId = "123")
-  expect_equal(target$username, "ron_smith")
+  target <- deploymentTarget(app_dir, appName = "foo")
+  expect_equal(target$username, "ron")
+})
+
+test_that("default title is the empty string", {
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
+
+  app_dir <- withr::local_tempdir()
+  file.create(file.path(app_dir, "app.R"))
+
+  target <- deploymentTarget(app_dir)
+  expect_equal(target$appTitle, "")
+})
+
+test_that("on first deploy only, title affects app name", {
+  mockr::local_mock(accounts = fakeAccounts("ron", "bar"))
+
+  app_dir <- withr::local_tempdir()
+  file.create(file.path(app_dir, "app.R"))
+
+  target <- deploymentTarget(app_dir, appTitle = "my title")
+  expect_equal(target$appName, "my_title")
+
+  saveDeployment(
+    app_dir,
+    name = "my_title",
+    title = "my title",
+    appId = "1",
+    bundleId = "",
+    url = ""
+  )
+  target <- deploymentTarget(app_dir, appTitle = "my new title")
+  expect_equal(target$appName, "my_title")
 })